add discipline crud endpoints

+++ server/src/controllers/studentController.ts
@@ -15,8 +15,12 @@ export const getStudents = async (req: Request, res: Response) => {
 
 export const getStudentById = async (req: Request, res: Response) => {
   const studentId = req.params.id;
+
   try {
     const studentById = await db.select().from(student).where(eq(student.id, Number(studentId)));
+    if (studentById.length === 0) {
+      return res.status(404).json({ message: "student doesn't exist" })
+    }
     return res.status(200).json(studentById);
   } catch (error) {
     return res.status(400).json({ message: "could not find student", error: error });
@@ -40,8 +44,12 @@ export const createStudent = async (req: Request, res: Response) => {
 
 export const deleteStudent = async (req: Request, res: Response) => {
   const studentId = req.params.id;
+  const studentById = await db.select().from(student).where(eq(student.id, Number(studentId)));
+  if (studentById.length === 0) {
+    return res.status(404).json({ message: "student doesn't exist" })
+  }
   try {
-    const studentById = await db.delete(student).where(eq(student.id, Number(studentId)));
+    const updatedStudent = await db.delete(student).where(eq(student.id, Number(studentId)));
     return res.status(200).json({ message: "student has been deleted" });
   } catch (error) {
     return res.status(400).json({ message: "could not delete student", error: error });
@@ -51,6 +59,10 @@ export const deleteStudent = async (req: Request, res: Response) => {
 export const updateStudent = async (req: Request, res: Response) => {
   const studentId = req.params.id;
   const { firstName, lastName } = req.body;
+  const studentById = await db.select().from(student).where(eq(student.id, Number(studentId)));
+  if (studentById.length === 0) {
+    return res.status(404).json({ message: "student doesn't exist" })
+  }
   try {
     const updatedStudent = await db.update(student).set({
       firstName: firstName,
